Polish: smooth tile arrival, WebGPU first, fresh frustum
- play.html creates the device with WebGPU first and WebGL2 as the
  fallback (WebGL2 only under ?xr=1). Under WebGL2 the unified gsplat
  path sorts on the CPU and reads each new tile's centres back with a
  synchronous readPixels, which is the stall on every arrival.
- WEBGL_LIMITS (4 M splats, 48 tiles, 2 in flight) for the CPU-sort
  path; WebGPU keeps LIMITS. The status line names the device.
- Arrivals are placed one per frame (TileStreamer.placeNext); a tile
  counts as in flight until it is placed.
- The traversal runs when the view or the in-flight set changes and
  every tenth frame otherwise; tile radius and centre are cached by
  key, anchor and sha.
- The culling frustum is built from the camera node's current
  transform rather than the engine's last-rendered one, a frame behind.
- nearClip 1 -> 0.3; the status line is written only when it changes.
- stream.spec turns KEEP_MS off through limits.keepMs; xr.spec asks the
  page for its device's budget instead of a stale constant.
- db/test 0093 and 0094: a smallint compared as one, and a ready ->
  verified shortcut that the state machine forbids, so db-test passes.

+++ db/test/0093_afailedpiecegoesbackatonce.sql
@@ -32,7 +32,7 @@ SELECT set_config('request.jwt.claims',
     json_build_object('sub', holder_id, 'role', 'player')::text, true) FROM ids;
 SELECT is(fail_atom(930001, 'no ground at 14/9301/9301'), 'ready',
     'the first failure puts it back in the pool');
-SELECT is((SELECT attempts FROM atom WHERE id = 930001), 1, 'and counts an attempt');
+SELECT is((SELECT attempts FROM atom WHERE id = 930001), 1::smallint, 'and counts an attempt');
 SELECT is((SELECT worker_id FROM atom WHERE id = 930001), null, 'in nobody''s hands');
 SELECT is((SELECT result ->> 'error' FROM atom WHERE id = 930001),
     'no ground at 14/9301/9301', 'with the reason recorded');
@@ -43,7 +43,7 @@ UPDATE atom SET state = 'claimed', worker_id = '00000000-0000-0000-0000-00000000
                 claimed_at = now(), attempts = 2 WHERE id = 930001;
 SELECT is(fail_atom(930001, 'still no ground'), 'failed',
     'three failures and the piece has given up');
-SELECT is((SELECT attempts FROM atom WHERE id = 930001), 3, 'all three counted');
+SELECT is((SELECT attempts FROM atom WHERE id = 930001), 3::smallint, 'all three counted');
 SELECT is(fail_atom(930002, 'x'), 'missing', 'an atom that does not exist is missing');
 
 SELECT * FROM finish();
